Back-end/requirements.txt
autopep8==2.0.1
blinker==1.6.1
cachelib==0.10.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
DateTime==5.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Mail==0.9.1
Flask-Session==0.4.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.1
psycopg2-binary==2.9.5
pycodestyle==2.10.0
PyJWT==2.6.0
PyMySQL==1.0.2
pytz==2023.2
requests==2.28.2
six==1.16.0
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
zope.interface==6.0